As discussed with the team, renamed the utils file to homeService.js since the funtions in this file are only used in the Home Component.
